update usestate for position component
